+++ Packages/com.espale.ui/Runtime/Scripts/Layers/UILayerManager.cs
@@ -1,14 +1,21 @@
 using System.Collections.Generic;
-using Espale.Utilities;
+using System.Linq;
+using UnityEngine;
 
 namespace Espale.UI.Layers
 {
-    public class UILayerManager : Singleton<UILayerManager>
+    // This class could actually be a static class but when we store it as a MonoBehavior, we can directly use it's prefab
+    // to call these methods from UI events.
+    public class UILayerManager : MonoBehaviour
     {
-        public static Stack<UILayer> activeLayers = new ();
+        public static readonly Stack<UILayer> activeLayers = new ();
 
-        public static void FocusLayer(UILayer layer) => GetInstance().FocusLayerFromInstance(layer);
-        private void FocusLayerFromInstance(UILayer layer)
+        /// <summary>
+        /// Focuses on the given layer. If there exists other active layer(s), makes those layers inactive if they are >= in
+        /// terms of layers, else makes them sleep.
+        /// </summary>
+        /// <param name="layer">Layer to focus to</param>
+        public static void FocusLayer(UILayer layer)
         {
             if (activeLayers.Count < layer.layerOrder)
             {
@@ -42,10 +49,41 @@ namespace Espale.UI.Layers
             // Focus the layer and push it to the stack.
             activeLayers.Push(layer);
             SwitchState(layer, UILayer.UILayerState.Focused);
+            
+            AdjustInteractivites();
         }
 
-        public static void HideLayer() => GetInstance().HideTopLayerFromInstance();
-        private void HideTopLayerFromInstance()
+        /// <summary>
+        /// Hides the given layer and every other layer above it, if any.
+        /// </summary>
+        /// <param name="layer">Layer to hide</param>
+        public static void HideLayer(UILayer layer)
+        {
+            if (!activeLayers.Contains(layer)) return;
+
+            while (activeLayers.Count > 0)
+            {
+                var topLayer = activeLayers.Peek(); // Cache the top layer.
+                HideTopLayer(); // Hide the top layer.
+                if (topLayer == layer) return; // If the hid top layer is the desired layer, terminate.
+            }
+        }
+
+        /// <summary>
+        /// If the given layer is not inactive, uses `UILayerManager.HideLayer()` method on the layer, else, it applies
+        /// the `UILayerManager.FocusLayer()` method.
+        /// </summary>
+        /// <param name="layer">Target layer</param>
+        public static void ToggleLayer(UILayer layer)
+        {
+            if (layer.state is not UILayer.UILayerState.Inactive) HideLayer(layer);
+            else FocusLayer(layer);
+        }
+        
+        /// <summary>
+        /// Hides the layer at the top of the active layers stack.
+        /// </summary>
+        public static void HideTopLayer()
         {
             if (activeLayers.Count <= 0) return;
             
@@ -60,6 +98,22 @@ namespace Espale.UI.Layers
             // If there is a layer left, focus it.
             if (activeLayers.Count > 0)
                 SwitchState(activeLayers.Peek(), UILayer.UILayerState.Focused);
+
+            AdjustInteractivites();
+        }
+
+        private static void AdjustInteractivites()
+        {
+            var layers = activeLayers.Reverse().Where(l => l != null).ToArray();
+            if (layers.Length < 2) return;
+            
+            for (var i = 0; i < layers.Length - 1; i++)
+            {
+                var layer = layers[i];
+                var layerAbove = layers[i + 1];
+                
+                SwitchState(layer, layer.state, layerAbove.blockLayersBelow);
+            }
         }
 
         /// <summary>
@@ -67,10 +121,11 @@ namespace Espale.UI.Layers
         /// </summary>
         /// <param name="layer">Layer to change the state of</param>
         /// <param name="state">Desired state for the layer</param>
-        private static void SwitchState(UILayer layer, UILayer.UILayerState state)
+        /// <param name="interactivityLimitedByAbove">Determines where the layers above are blocking the interactivity</param>
+        private static void SwitchState(UILayer layer, UILayer.UILayerState state, bool interactivityLimitedByAbove=true)
         {
             if (layer) 
-                layer.SwitchState(state);
+                layer.SwitchState(state, interactivityLimitedByAbove);
         }
     }
 }